perf(runtime): the largest term in scr_cyc_alloc was the stack frame it kept for a call it did not have to make
scr_cyc_alloc is 20.62% of closure-churn at 60 instructions per call.
Disassembled (tests/perf/cycalloc/isa.mjs, exact counts on 800,031 calls),
16 of the 60 are the stack frame: seven pushes, seven pops, an rsp
alignment and the frame pointer. The allocation itself -- the round-up,
the range check, the pool pop and the header stores -- is 30.

The frame existed for two calls, and both stop being calls here.

The memset was the first. The body zeroed the payload FIRST and stamped
the header after, so `trace`, `free_fn` and `phys` all had to survive it
and clang put them in callee-saved registers. Stamping the header first
leaves nothing live across the memset -- and memset RETURNS ITS
DESTINATION, which is exactly this function's return value, so
`return memset(h + 1, 0, phys - sizeof *h)` is the old `return h + 1`
with the zeroing folded into a tail call.

calloc was the second. It runs 5 times in 800,031 and it cost the hot
path five callee-saved registers to be reachable. The whole miss is out
of line now, taking the same three arguments so it is a tail call too.

Two more terms are one line each. The size-class index cost 5
instructions because `int c` makes `p->head[c]` a signed index and clang
fuses the sign extension into shl/movabs/add/sar; unsigned it is one
`shr` with the -1 folded into the addressing mode, and it now CSEs with
the `phys / SCR_POOL_GRAIN` the header's `blk` field needs. `blk` itself
cost 5 because it recomputed that quotient and re-tested a range the
pooled arm had already proved -- it is an argument now, so each arm
passes what it knows.

  self Ir/call   60 -> 33     stack frame 16 -> 3, class index 5 -> 1,
                              blk 5 -> 1

The memset is unchanged at 12 instructions inside libc, and is not in
either figure: zeroing 64 bytes is real work and no inlining removes it.

+++ packages/runtime/src/scr_runtime.h
@@ -342,7 +342,17 @@ static inline void *scr_pool_take(ScrPool *p, size_t n) {
 #else
   size_t r = scr_pool_bytes(n);
   if (r == 0 || r > SCR_POOL_MAX) return NULL;
-  int c = (int)(r / SCR_POOL_GRAIN) - 1;
+  /* size_t rather than int, and that is worth five instructions here.
+   * `int c` makes `p->head[c]` a SIGNED index, so clang must sign-extend
+   * it before it can scale by 8 -- and because `r` is itself derived from
+   * a shift it fuses the two into `shl $0x1d / movabs / add / sar $0x20`,
+   * four instructions plus the shift, on the hottest path in the runtime.
+   * Unsigned, the same index is `shr $0x3` and a decrement folded into the
+   * addressing mode, and it CSEs with the `phys / SCR_POOL_GRAIN` the
+   * caller needs for ScrCycHdr::blk. `r` is non-zero and a multiple of the
+   * grain by the two lines above, so the subtraction cannot wrap.
+   * Read out of scr_cyc_alloc's disassembly, tests/perf/cycalloc/isa.mjs. */
+  size_t c = r / SCR_POOL_GRAIN - 1u;
   void *b = p->head[c];
   if (!b) return NULL;
   /* the free-list link lives in the block's first word; every class is
@@ -363,7 +373,7 @@ static inline bool scr_pool_give(ScrPool *p, void *b, size_t n) {
 #else
   size_t r = scr_pool_bytes(n);
   if (r == 0 || r > SCR_POOL_MAX) return false;
-  int c = (int)(r / SCR_POOL_GRAIN) - 1;
+  size_t c = r / SCR_POOL_GRAIN - 1u;  /* unsigned; see scr_pool_take */
   if (p->n[c] >= SCR_POOL_DEPTH) return false;
   __builtin_memcpy(b, &p->head[c], sizeof(void *));
   p->head[c] = b;
